Fixed warnings

+++ Sources/OAuth2Client/NXOAuth2PostBodyStream.m
@@ -30,7 +30,7 @@
 {
     self = [self init];
     if (self) {
-        srandom(time(NULL));
+        srandom((unsigned int)time(NULL));
         boundary = [[NSString alloc] initWithFormat:@"------------nx-oauth2%d", rand()];
         numBytesTotal = 0;
         streamIndex = 0;
@@ -88,7 +88,7 @@
             NSData *delimiterData = [delimiter dataUsingEncoding:NSUTF8StringEncoding];
             NSData *contentHeaderData = [[part contentHeaders] dataUsingEncoding:NSUTF8StringEncoding];
             
-            int dataLength = delimiterData.length + contentHeaderData.length;
+            NSUInteger dataLength = delimiterData.length + contentHeaderData.length;
             NSMutableData *headerData = [NSMutableData dataWithCapacity: dataLength];
             [headerData appendData:delimiterData];
             [headerData appendData:contentHeaderData];
@@ -146,7 +146,7 @@
     if (currentStream == nil)
         return 0;
     
-    int result = [currentStream read:buffer maxLength:len];
+    NSInteger result = [currentStream read:buffer maxLength:len];
     
     if (result == 0) {
         if (streamIndex < contentStreams.count - 1) {
